Share code for the podman and kubernetes installations
